xroads_django/requirements.txt
asgiref==3.2.10
attrs==19.3.0
autopep8==1.5.3
coverage==5.2
dj-rest-auth==1.1.0
Django==3.0.8
djangorestframework==3.11.0
more-itertools==8.4.0
packaging==20.4
Pillow==6.2.1
pip-autoremove==0.9.1
pkg-resources==0.0.0
pluggy==0.13.1
postgres==2.2.2
psycopg2==2.8.3
psycopg2-binary==2.8.3
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-randomly==3.4.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wcwidth==0.2.5
